Move Streams/Zip.hs to Internal/Data/Stream
